requirements.txt
fastapi
uvicorn
clickhouse-connect
python-dotenv
python-jose[cryptography]
pydantic
faker
httpx
pytest